requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
pydantic
pydantic-settings
python-multipart
email-validator
httpx
google-genai
chromadb
groq
# Phase 1: Perplexity-class Intelligence
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-core>=0.3.0

# Voice Integration
faster-whisper==1.0.2
git+https://github.com/huggingface/parler-tts.git
librosa==0.10.0
soundfile==0.12.1
torch>=2.1.0
pyaudio==0.2.14
numpy<2.0.0
# Fine-tuning
peft
bitsandbytes
datasets
accelerate
